feat: gatling conf

+++ tests/gatling/entrypoint.sh
@@ -2,31 +2,54 @@
 
 ln -sf /usr/share/zoneinfo/America/Sao_Paulo /etc/localtime
 
-if [ ! -d "bundle/bin" ]; then
-    cd bundle
-
-    echo "Downloading Gatling bundle..."
-    wget https://repo1.maven.org/maven2/io/gatling/highcharts/gatling-charts-highcharts-bundle/3.9.5/gatling-charts-highcharts-bundle-3.9.5-bundle.zip
-
-    echo "Unzip Gatling bundle..."
-    unzip gatling-charts-highcharts-bundle-3.9.5-bundle.zip
-
-    echo "Remove zip bundle..."
-    rm -rf gatling-charts-highcharts-bundle-3.9.5-bundle.zip
-
-    cd ..
-
-    echo "Populate folder bundle..."
-    mv bundle/gatling-charts-highcharts-bundle-3.9.5/* bundle
-
-    echo "Remove original gatling folder..."
-    chmod -R 777 bundle
-    rm -rf bundle/gatling-charts-highcharts-bundle-3.9.5
-
-    sleep 5
+if [ ! -e /entrypoint ]; then
+    ln -s /usr/src/app/entrypoint.sh /entrypoint
 fi
 
-echo "EXECUTE GatlingTest..."
-$(pwd)/bundle/bin/gatling.sh -rm local -rd GatlingTest-$(exec date "+%m/%d/%Y-%H:%M:%S") -sf $(pwd)/APISimulation -rf $(pwd)/results APISimulation.Alunos
+if [ "$1" = "run-test" ]; then
 
-nc -l -p 8080
+    if [ ! -d "bundle/bin" ]; then
+
+        cd bundle
+
+        echo "Downloading Gatling bundle..."
+        wget  https://repo1.maven.org/maven2/io/gatling/highcharts/gatling-charts-highcharts-bundle/3.9.5/gatling-charts-highcharts-bundle-3.9.5-bundle.zip
+
+        echo "Unzip Gatling bundle..."
+        unzip gatling-charts-highcharts-bundle-3.9.5-bundle.zip
+
+        echo "Remove zip bundle..."
+        rm -rf gatling-charts-highcharts-bundle-3.9.5-bundle.zip
+
+        cd ..
+
+        echo "Populate folder bundle..."
+        mv bundle/gatling-charts-highcharts-bundle-3.9.5/* bundle
+
+        echo "Remove original gatling folder..."
+        chmod -R 777 bundle
+        rm -rf bundle/gatling-charts-highcharts-bundle-3.9.5
+
+    fi
+
+    echo "EXECUTE GatlingTest..."
+    description=LoadTest::$API_NAME::v$API_TAG_VERSION::$(exec date "+%m/%d/%Y-%H:%M:%S")::America/Sao_Paulo
+    $(pwd)/bundle/bin/gatling.sh -rm local -rd $description -sf $(pwd)/APISimulation -rf $(pwd)/results/history APISimulation.Alunos
+    echo "Verify Test Gatling Results folder for all tests"
+fi
+
+echo "Remove last load test data"
+rm -rf ./results/latest/*
+touch ./results/latest/.touch
+
+echo "Add New load test data"
+new_latest=$(ls -td ./results/history/*/ | head -n 1)
+cp -r $new_latest/* ./results/latest/
+
+python3_pid=$(pgrep -f "python3 -m http.server $TEST_GATLING_PORT")
+if [ ! -n "$python3_pid" ]; then
+    echo "Run test last result server"
+    python3 -m http.server $TEST_GATLING_PORT --directory ./results/latest/
+fi
+
+
